+++ scilab/modules/slint/src/cpp/DeprecatedChecker.cpp
@@ -95,6 +95,7 @@ std::unordered_map<std::wstring, std::wstring> DeprecatedChecker::initDep()
     // Scilab 6.0.x => 6.1.0
     map.emplace(L"_d", L"_");
     map.emplace(L"dgettext", L"gettext");
+    map.emplace(L"datatipToggle", L"datatipManagerMode");
     map.emplace(L"denom", L".den");
     map.emplace(L"lstsize", L"size");
     map.emplace(L"nanmin", L"min");